requirements.txt
# Core dependencies
requests>=2.31.0
aiohttp>=3.8.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
feedparser>=6.0.0

# Data processing
pandas>=2.0.0
python-dateutil>=2.8.0

# Web framework (optional)
flask>=2.3.0
flask-cors>=4.0.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
colorama>=0.4.6
rich>=13.0.0

# Scheduling
schedule>=1.2.0

# Async support
aiofiles>=23.0.0

# News processing
newspaper3k>=0.2.8
textblob>=0.17.1


# Optional: for web scraping with browser automation
# selenium>=4.15.0
# webdriver-manager>=4.0.0